fix: Chunk._compute_semantic_signature returns Python floats
The normalized branch returned [x / norm for x in signature] where
signature is np.float32 ndarray and norm is np.float32 scalar, so list
elements stayed as np.float32 instead of Python floats. The type contract
(list[float]) and test_returns_list_of_128_floats both assumed Python
floats. Adds float() cast on the normalized path to match the existing
cast on the zero-norm path.

Prevents downstream JSON-serialization surprises — np.float32 is not
JSON-serializable by default.

+++ stele_context/chunkers/base.py
@@ -241,7 +241,7 @@ class Chunk:
         # Normalize to unit vector
         norm = np.linalg.norm(signature)
         if norm > 0:
-            return [x / norm for x in signature]
+            return [float(x / norm) for x in signature]
 
         return [float(x) for x in signature]
